feat(hipblaslt): extend GPU correctness reference to bf16, fp8, bf8, f64
Widen the opt-in GPU correctness reference (--check_ref gpu|both) beyond
f32/f16:

- Support bf16, OCP fp8 (e4m3), OCP bf8 (e5m2), and f64 inputs; C/D in
  {f32, f16, bf16, f64}. fp8/bf8 are input-only.
- Template the reference GEMM on the accumulate type: float for the
  compute-32F types, double for f64 (compute 64F), selected by the output
  type.
- Add a double-precision device compare and a 4-ULP double unit check for
  f64; widen the reference alpha/beta to double.
- Decode fp8/bf8 inputs via a device-portable software conversion so the
  kernel builds on architectures without native OCP fp8 conversion.

The support gate requires a_type == b_type, pairs compute 64F with all-f64
operands, and rejects FNUZ fp8/bf8, MX, scaling, and epilogues.

+++ projects/hipblaslt/clients/common/include/reference_device.hpp
@@ -13,26 +13,30 @@
  *        lives in gpu_compare.hpp). Used only for testing/benchmark verification;
  *        not part of the GPU library.
  *
- * One thread per output element, float accumulate, implemented independently of
- * TensileLite so it can catch kernel bugs. Selected via the `check_ref` argument
- * / `--check_ref` bench flag.
+ * One thread per output element, implemented independently of TensileLite so it
+ * can catch kernel bugs. f64 accumulates in double; every other supported type
+ * accumulates in float. Selected via the `check_ref` argument / `--check_ref`
+ * bench flag.
  */
 
 /// True when `arg` describes a matmul the GPU reference path currently supports
-/// (plain f32/f16 GEMM, default epilogue, no scaling/bias/aux, strided batch).
-/// On false, `reason` is filled with the first unsupported feature encountered.
+/// (plain GEMM with f32/f16/bf16/OCP-fp8/OCP-bf8 inputs and f32/f16/bf16/f64
+/// C/D, compute 32F or all-f64 compute 64F, default epilogue, no
+/// scaling/bias/aux, strided batch). On false, `reason` is filled with the first
+/// unsupported feature encountered.
 bool gpu_ref_supported(const Arguments& arg, std::string& reason);
 
-/// Compute D_gold = alpha * op(A) * op(B) + beta * C on the device, accumulating
-/// in float. All pointers are device pointers. Column-major, with the same
+/// Compute D_gold = alpha * op(A) * op(B) + beta * C on the device. f64 (tD ==
+/// HIP_R_64F) accumulates in double; every other type accumulates in float. All
+/// pointers are device pointers. Column-major, with the same
 /// transpose/leading-dim/batch-stride conventions as cblas_gemm().
 void run_reference_gemm_device(bool        transA_is_n,
                                bool        transB_is_n,
                                int64_t     M,
                                int64_t     N,
                                int64_t     K,
-                               float       alpha,
-                               float       beta,
+                               double      alpha,
+                               double      beta,
                                const void* dA,
                                hipDataType tA,
                                int64_t     lda,
